test: structure test by features
